scraping-hub/pyproject.toml
[tool.poetry]
name = "scraping-hub"
version = "0.1.0"
description = ""
authors = ["yamamototkm <test@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.40"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
slack-sdk = "^3.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
